performance improvement
Serious improvement: don't store empty intervals in a History. This
makes a huge difference, because after several rotations, there are tons
of empty intervals lying around whose only purpose was to turn off a
signal. Notice that we have to propogate these empty signals (since it
may be important that they override non-empty signals at other nodes),
but once they're propogated, they can be discarded.

A more minor improvement was to switch from Rational to Double. The (+.)
function was taking about 30% of the entire runtime, but after switching
to hardware floats, this dropped to practically 0%, so we saved 30% on
runtime by just this change.

This commit also includes some commented-out benchmark code that can be
used when profiling these kinds of changes.